Able to pan and enable laser with remote
NecRemote is currently full of debug output. I'm definitely using the
decoder incorrectly in non-blocking mode. As a workaround I'm
reading the raw values out of pulseio and waiting for something that
looks like it could possibly be the start of a Nec code then using the
decoder in blocking mode since I know data is available.

This is likely good enough for the purposes of this project.